REFACTOR: moved all material into directory - Grouping material into the same location
