+++ actions/projects/getProjectBySlug.ts
@@ -1,4 +1,4 @@
-import ProjectInterface from "@/interfaces/ProjectInterface";
+import ProjectInterface from "@/interfaces/material/ProjectInterface";
 
 /**
  * Gets the project with the given slug (unique identifier).
