+++ components/CertificateItem/CertificateItem.tsx
@@ -3,7 +3,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from "@/components/shadcn/ui/tooltip";
-import CertificateInterface from "@/interfaces/CertificateInterface";
+import CertificateInterface from "@/interfaces/material/CertificateInterface";
 import Image from "next/image";
 import Link from "next/link";
 import React from "react";
